Android Studio formatting for whole project, forgot to do it on the previous commit

+++ lib/filters/entry/filters_entry.dart
@@ -151,8 +151,7 @@ class FilterEntryState extends State<FilterEntry> {
     FaceLandmarkType.values.forEach((landmarkType) {
       var tempFile = getAppFile(getLandmarkFilename('temp', landmarkType));
       var newPath = getInternalFilename(getLandmarkFilename(model.entityBeingEdited.name, landmarkType));
-      if (tempFile.existsSync())
-        tempFile.renameSync(newPath);
+      if (tempFile.existsSync()) tempFile.renameSync(newPath);
     });
 
     // Update entity
